+++ lib/pin_code_fields.dart
@@ -16,3 +16,4 @@ part 'src/models/animation_type.dart';
 part 'src/models/dialog_config.dart';
 part 'src/models/pin_theme.dart';
 part 'src/pin_code_fields.dart';
+part 'src/gradiented.dart';